+++ common/src/main/java/dev/mayaqq/estrogen/utils/client/ItemModelBufferer.java
@@ -4,64 +4,13 @@ import com.jozufozu.flywheel.core.model.BlockModel;
 import com.mojang.blaze3d.vertex.*;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.LightTexture;
-import net.minecraft.client.renderer.MultiBufferSource;
-import net.minecraft.client.renderer.RenderType;
-import net.minecraft.client.renderer.block.model.BakedQuad;
 import net.minecraft.client.renderer.entity.ItemRenderer;
 import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.world.item.ItemDisplayContext;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
-import java.util.Set;
 
 public class ItemModelBufferer {
-
-    private static final VertexConsumer EMPTY_CONSUMER = new VertexConsumer() {
-        @Override
-        public VertexConsumer vertex(double x, double y, double z) {
-            return this;
-        }
-
-        @Override
-        public VertexConsumer color(int red, int green, int blue, int alpha) {
-            return this;
-        }
-
-        @Override
-        public VertexConsumer uv(float u, float v) {
-            return this;
-        }
-
-        @Override
-        public VertexConsumer overlayCoords(int u, int v) {
-            return this;
-        }
-
-        @Override
-        public VertexConsumer uv2(int u, int v) {
-            return this;
-        }
-
-        @Override
-        public VertexConsumer normal(float x, float y, float z) {
-            return this;
-        }
-
-        @Override
-        public void endVertex() {}
-
-        @Override
-        public void defaultColor(int defaultR, int defaultG, int defaultB, int defaultA) {}
-
-        @Override
-        public void unsetDefaultColor() {}
-
-        @Override
-        public void putBulkData(PoseStack.Pose poseEntry, BakedQuad quad, float red, float green, float blue, int combinedLight, int combinedOverlay) {}
-
-        @Override
-        public void putBulkData(PoseStack.Pose poseEntry, BakedQuad quad, float[] colorMuls, float red, float green, float blue, int[] combinedLights, int combinedOverlay, boolean mulColor) {}
-    };
 
     private static final ThreadLocal<ThreadLocalObjects> LOCAL_OBJECTS = ThreadLocal.withInitial(ThreadLocalObjects::new);
 
@@ -106,22 +55,4 @@ public class ItemModelBufferer {
 
     }
 
-    // This fixes visual bugs as enchantment glint isn't currently supported by flywheel
-    private record GlintFilteringBufferSource(VertexConsumer wrapped) implements MultiBufferSource {
-
-        private static final Set<RenderType> GLINT_TYPES = Set.of(
-            RenderType.glint(),
-            RenderType.glintDirect(),
-            RenderType.glintTranslucent(),
-            RenderType.entityGlint(),
-            RenderType.entityGlintDirect(),
-            RenderType.armorGlint(),
-            RenderType.armorEntityGlint()
-        );
-
-        @Override
-        public VertexConsumer getBuffer(RenderType renderType) {
-            return GLINT_TYPES.contains(renderType) ? EMPTY_CONSUMER : wrapped;
-        }
-    }
 }
